# docs/api/receipt-parser.openapi.yaml
openapi: 3.1.0
info:
  title: Mekasa API — Receipt Intelligence (Gemini parser)
  version: 0.8.0-draft
  summary: Receipt upload, Gemini parse jobs, line-item resolution and confirmation.
  description: |
    Contract for the receipt-parser feature (spec REQ-RCP-001 … REQ-RCP-021;
    design `docs/design/gemini-receipt-parser.md`). Draft for review — the
    FastAPI implementation in phase 2 must generate an `/openapi.json` whose
    paths and schemas match this file (a phase-2 CI check diffs them).

    Conventions (unchanged from the rest of the Mekasa API):
    * Bearer Firebase ID token; `Bearer test:<uid>` when `ALLOW_TEST_AUTH=true`.
    * Every household route checks membership → `403 Forbidden` for
      non-members, `404` for unknown household/receipt/line.
    * Timestamps are RFC 3339 UTC. Money is a JSON number (USD in v1).
    * Error body: `{"detail": "<snake_case_code>"}`.
servers:
  - url: https://mekasa-api-934775015882.us-central1.run.app
    description: dev (Cloud Run)
  - url: http://localhost:8080
    description: local
security:
  - bearerAuth: []
...
paths:
...
  /v1/households/{household_id}/receipts/{receipt_id}/line-items/{line_item_id}/capture:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - $ref: '#/components/parameters/ReceiptId'
      - $ref: '#/components/parameters/LineItemId'
    post:
      tags: [line-items]
      operationId: captureLineItemProduct
      summary: Scan the product's barcode (and optionally photograph it) when no UPC was discovered
      description: |
        Satisfies: REQ-RCP-020 AC2–AC5, REQ-RCP-013 · Spec version: 1.0

        The fallback (usually at home while unpacking) for lines with `barcode: null`
        (`capture_available: true`). The client scans the package barcode with
        the existing camera scanner and may upload a photo first
        (`POST …/product-photos`).

        - UPC already in `products` → the line is linked (`match_method: user`,
          `resolution_status: confirmed`), confirmation counted once per
          household; a `photo_id` becomes the product image only when the
          product has none (otherwise it is offered as a correction).
        - UPC unknown → a new `products` row keyed by the UPC is created with
          `source: user_scan`, `confidence_score: 0.9`, attributes from the
          request falling back to the line's `name`/`category`, `image_url`
          from `photo_id`; an `enrichment_jobs` row is queued.
        - Line previously on an `llm:` product → re-keyed in the same
          transaction (REQ-RCP-010 AC2: `superseded_by`, alias repoint).

        Always writes a `scan_events` entry (`context: receipt_capture`,
        `outcome: found | created`) correlated with this receipt and line.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCaptureRequest' }
            examples:
              scanOnly: { value: { upc: "041220576037" } }
              scanAndPhoto:
                value:
                  upc: "041220576037"
                  photo_id: "6f1c2d4e-8a3b-4c5d-9e7f-0a1b2c3d4e5f"
                  name: "H-E-B Organics Cauliflower"
                  category: Produce
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductCaptureResponse' }
        '400':
          description: '`invalid_upc`, `invalid_plu`, `exactly_one_code_required`, `unknown_photo_id`'
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409':
          description: '`receipt_confirmed`'
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
...
  /v1/households/{household_id}/inventory/{item_id}/capture:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: item_id
        in: path
        required: true
        schema: { type: string }
    post:
      tags: [inventory]
      operationId: captureInventoryItemProduct
      summary: Scan + photograph an inventory item that has no product/barcode
      description: |
        Satisfies: REQ-RCP-020 AC6 · Spec version: 1.0

        Same body and catalog semantics as the line-item `capture`; additionally
        sets `inventory_items.barcode`, `product_id`, `image_url`. Writes a
        `scan_events` entry with `context: inventory_capture`.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ProductCaptureRequest' }
      responses:
        '200':
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProductCaptureResponse' }
        '400':
          description: '`invalid_upc`, `invalid_plu`, `exactly_one_code_required`, `unknown_photo_id`'
          content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    cloudTasksOidc:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC (Cloud Tasks)
  parameters:
    HouseholdId:
      name: household_id
      in: path
      required: true
      schema: { type: string }
    ReceiptId:
      name: receipt_id
      in: path
      required: true
      schema: { type: string }
    LineItemId:
      name: line_item_id
      in: path
      required: true
      schema: { type: string }
...
  responses:
...
    Forbidden:
      description: Caller is not a member of the household (or lacks the Owner role where required).
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
    NotFound:
      description: Household, receipt, line item, or product not found.
      content: { application/json: { schema: { $ref: '#/components/schemas/Error' } } }
  schemas:
    Error:
      type: object
      required: [detail]
      properties:
        detail: { type: string, examples: [image_or_text_required] }
...
    ResolutionStatus:
      type: string
      enum: [auto_matched, needs_confirmation, unmatched, confirmed, rejected, skipped]
    MatchMethod:
      type: string
      enum: [exact_upc, scan_correlation, alias, fuzzy, enrichment, user]
    ProductSource:
      type: string
      enum: [user_scan, store_api, gs1_registry, llm_ocr]
      description: '`store_api` = official retailer API (Kroger/Walmart.io); `gs1_registry` = UPC verified against GS1. Scraped pages are never a source.'
    ProductStatus:
      type: string
      enum: [unverified, pending, verified]
...
    LineItem:
      type: object
      description: |
        Superset of the legacy `ReceiptLineItem`: `name`, `category`,
        `quantity`, `price_paid`, `barcode`, `image_url`, `identified` keep
        their old meaning so existing confirm-haul screens keep working.
      required: [id, line_no, raw_text, name, category, quantity, qty, price_paid, image_url, identified, extraction_confidence, matched_product_id, resolution_status, candidate_product_ids]
      properties:
        id: { type: string }
        line_no: { type: integer, minimum: 1 }
        raw_text: { type: string }
        name:
          type: string
          description: Cleaned description (legacy field name).
        category: { type: string }
        quantity:
          type: integer
          minimum: 1
          description: |
            Integer quantity for inventory. `1` for weighed lines (`unit` lb/oz/kg/g:
            one bunch or bag as sold; the weight stays in `qty`); otherwise `qty`
            rounded, min 1 (REQ-RCP-006 AC5).
        qty: { type: number, exclusiveMinimum: 0 }
        unit: { type: [string, 'null'] }
        price_paid:
          type: [number, 'null']
          description: Extended price after discounts (legacy field name).
        unit_price: { type: [number, 'null'] }
        discount: { type: [number, 'null'] }
        printed_code: { type: [string, 'null'] }
        barcode:
          type: [string, 'null']
          description: UPC of the matched product, when known.
        brand: { type: [string, 'null'] }
        unit_size: { type: [string, 'null'] }
        image_url:
          type: string
          description: |
            User photo (`/v1/product-photos/{photo_id}`) when the line has one,
            else the product image, else a category placeholder.
        image_source: { $ref: '#/components/schemas/ImageSource' }
        photo_id:
          type: [string, 'null']
          format: uuid
          description: Set when the household attached its own photo to this line (REQ-RCP-019 AC4).
        capture_available:
          type: boolean
          description: |
            True when `barcode` is null or `resolution_status` is
            `needs_review`/`unresolved`: the client offers "Scan product"
            (REQ-RCP-020 AC1). False when `no_barcode` is true.
        no_barcode:
          type: boolean
          description: |
            The household marked this receipt text (per store chain) as having
            no barcode (REQ-RCP-020 AC10); the client treats the line as ready
            and shows plain "No barcode" text (no chip) instead of "Scan →"
            (REQ-RCP-020 AC11).
        user_edited_fields:
          type: array
          items:
            type: string
            enum: [name, brand, category, unit_size, qty, unit, quantity, price_paid, unit_price, photo_id]
          description: Fields the household overrode via PATCH (REQ-RCP-019 AC1).
        identified:
          type: boolean
          description: '`resolution_status` in {auto_matched, confirmed}.'
        extraction_confidence: { type: number, minimum: 0, maximum: 1 }
        matched_product_id: { type: [string, 'null'] }
        match_method:
          oneOf: [{ $ref: '#/components/schemas/MatchMethod' }, { type: 'null' }]
        match_confidence: { type: [number, 'null'], minimum: 0, maximum: 1 }
        resolution_status: { $ref: '#/components/schemas/ResolutionStatus' }
        candidate_product_ids:
          type: array
          maxItems: 5
          items: { type: string }
        conflict:
          oneOf: [{ $ref: '#/components/schemas/ProductConflict' }, { type: 'null' }]
        inventory_item_id: { type: [string, 'null'] }
        purchase_event_id: { type: [string, 'null'] }
...
    ProductCaptureRequest:
      type: object
      additionalProperties: false
      oneOf:
        - required: [upc]
          not: { required: [plu_code] }
        - required: [plu_code]
          not: { required: [upc] }
      description: |
        Product capture (REQ-RCP-020 AC2/AC7), usually at home while unpacking: the scanned barcode, or the
        PLU typed from a produce sticker, plus an optional photo and attribute
        overrides. Missing attributes fall back to the line / inventory item
        (`name`, `category`) when a product has to be created. A `plu_code`
        capture links to or creates the shared product `plu:<code>`.
      properties:
        upc: { type: string, pattern: '^[0-9]{8,14}$' }
        plu_code: { type: string, pattern: '^[0-9]{4,5}$' }
        photo_id: { type: [string, 'null'], format: uuid }
        name: { type: [string, 'null'], minLength: 1, maxLength: 120 }
        brand: { type: [string, 'null'], maxLength: 80 }
        category: { type: [string, 'null'], minLength: 1, maxLength: 60 }
        unit_size: { type: [string, 'null'], maxLength: 40 }
        receipt_text:
          type: [string, 'null']
          maxLength: 200
          description: 'Inventory variant: the receipt line text the item came from; upserted as an alias so the next scan matches (REQ-RCP-020 AC6).'
        store_chain_id:
          type: [string, 'null']
          maxLength: 40
          description: '`store_chain_id` from the scan response; unknown chains fall back to `unknown`.'
    ProductCaptureResponse:
      type: object
      required: [outcome, product, scan_event_id, confirmation_counted]
      properties:
        outcome:
          type: string
          enum: [linked, created, rekeyed]
          description: '`linked` = UPC already known; `created` = new `user_scan` product; `rekeyed` = previous `llm:` product superseded.'
        product: { $ref: '#/components/schemas/Product' }
        line_item:
          oneOf: [{ $ref: '#/components/schemas/LineItem' }, { type: 'null' }]
          description: Present for the line-item variant.
        inventory_item_id:
          type: [string, 'null']
          description: Present for the inventory variant.
        inventory_item:
          type: [object, 'null']
          description: |
            Inventory variant only: the updated item (`barcode` is set for UPC
            captures only, `product_id` is the UPC or `plu:<code>`, `image_url`
            is the capture photo when one was sent).
        scan_event_id: { type: string }
        confirmation_counted: { type: boolean }
        enrichment_job_id:
          type: [string, 'null']
          description: Queued for `created`/`rekeyed` so store/registry adapters can fill gaps (REQ-RCP-020 AC4).
        photo_applied_as:
          type: string
          enum: [product_image, line_image, correction_proposed, none]
          description: |
            Where the submitted `photo_id` ended up: the product had no image
            (`product_image`), only this household's line (`line_image`), or a
            `product_conflicts` record because the product is verified
            (`correction_proposed`).
        conflict:
          oneOf: [{ $ref: '#/components/schemas/ProductConflict' }, { type: 'null' }]
...
    ImageSource:
      type: string
      enum: [user_photo, store_api, openfoodfacts, gs1_registry, placeholder]
      description: Provenance of `image_url`; `user_photo` means a `/v1/product-photos/{photo_id}` URL.
...
    Product:
      type: object
      required: [id, upc, store_chain_id, name, category, source, confidence_score, confirmation_count, status]
      properties:
        id: { type: string }
        code_kind: { type: string, enum: [upc, plu, llm], description: '`plu` rows are keyed `plu:<code>` and filed under store chain `unknown` (shared across chains).' }
        upc: { type: [string, 'null'] }
        plu_code: { type: [string, 'null'], pattern: '^[0-9]{4,5}$' }
        store_chain_id: { type: string }
        name: { type: string }
        brand: { type: [string, 'null'] }
        category: { type: string }
        unit_size: { type: [string, 'null'] }
        image_url: { type: [string, 'null'] }
        image_source:
          oneOf: [{ $ref: '#/components/schemas/ImageSource' }, { type: 'null' }]
        source: { $ref: '#/components/schemas/ProductSource' }
        confidence_score: { type: number, minimum: 0, maximum: 1 }
        confirmation_count: { type: integer, minimum: 0 }
        status: { $ref: '#/components/schemas/ProductStatus' }
        superseded_by: { type: [string, 'null'] }
    ProductConflict:
      type: object
      required: [id, product_id, field, verified_value, observed_value, status]
      properties:
        id: { type: string }
        product_id: { type: string }
        field: { type: string, enum: [name, brand, unit_size, category, upc, image_url] }
        verified_value: { type: string }
        observed_value: { type: string }
        status: { type: string, enum: [open, dismissed, accepted] }
...